+++ gcs/src/main/java/com/google/cloud/hadoop/fs/gcs/RefreshTokenAuth2Provider.java
@@ -1,38 +1,31 @@
 package com.google.cloud.hadoop.fs.gcs;
 
-import com.fasterxml.jackson.core.JsonFactory;
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.core.JsonToken;
+import com.google.api.client.auth.oauth2.ClientParametersAuthentication;
+import com.google.api.client.auth.oauth2.RefreshTokenRequest;
+import com.google.api.client.auth.oauth2.TokenRequest;
+import com.google.api.client.auth.oauth2.TokenResponse;
+import com.google.api.client.http.GenericUrl;
+import com.google.api.client.http.HttpTransport;
+import com.google.api.client.json.JsonFactory;
+import com.google.api.client.json.jackson2.JacksonFactory;
 import com.google.cloud.hadoop.gcsio.GoogleCloudStorageOptions;
 import com.google.cloud.hadoop.util.AccessTokenProvider;
+import com.google.cloud.hadoop.util.HttpTransportFactory;
 import com.google.cloud.hadoop.util.RedactedString;
 import com.google.common.flogger.GoogleLogger;
-import com.google.common.io.BaseEncoding;
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.InetSocketAddress;
-import java.net.Proxy;
-import java.net.URL;
-import java.nio.charset.StandardCharsets;
 import java.time.Instant;
 import java.time.ZoneId;
 import java.time.format.DateTimeFormatter;
 import org.apache.hadoop.conf.Configuration;
-import org.apache.http.HttpHeaders;
 
 public class RefreshTokenAuth2Provider implements AccessTokenProvider {
   private static final GoogleLogger logger = GoogleLogger.forEnclosingClass();
-  private static DateTimeFormatter dateFormat =
-      DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss.SSS").withZone(ZoneId.systemDefault());
-
-  private static final int CONNECT_TIMEOUT = 30 * 1000;
-  private static final int READ_TIMEOUT = 30 * 1000;
+  private static final JsonFactory JSON_FACTORY = JacksonFactory.getDefaultInstance();
 
   private static final AccessToken EXPIRED_TOKEN = new AccessToken("", -1L);
-
+  private static DateTimeFormatter dateFormat =
+      DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss.SSS").withZone(ZoneId.systemDefault());
   private Configuration config;
   private AccessToken accessToken = EXPIRED_TOKEN;
 
@@ -53,7 +46,7 @@ public class RefreshTokenAuth2Provider implements AccessTokenProvider {
 
     GoogleCloudStorageOptions gcsOptions =
         GoogleHadoopFileSystemConfiguration.getGcsOptionsBuilder(this.config).build();
-    String tokenEndpoint = gcsOptions.getTokenEndpoint();
+    String tokenServerUrl = gcsOptions.getTokenServerUrl();
     RedactedString refreshToken = gcsOptions.getRefreshToken();
     String clientId = gcsOptions.getClientId();
     RedactedString clientSecret = gcsOptions.getClientSecret();
@@ -62,14 +55,18 @@ public class RefreshTokenAuth2Provider implements AccessTokenProvider {
     RedactedString proxyPassword = gcsOptions.getProxyPassword();
 
     logger.atFine().log(
-        "Refresh token calling endpoint '" + tokenEndpoint + "' with client id '" + clientId + "'");
+        "Refresh token calling endpoint '"
+            + tokenServerUrl
+            + "' with client id '"
+            + clientId
+            + "'");
     logger.atFine().log(
         "Proxy setup: '" + proxyAddress + "' with username = '" + proxyUsername + "'");
 
     try {
       this.accessToken =
           getAccessToken(
-              tokenEndpoint,
+              tokenServerUrl,
               clientId,
               clientSecret,
               refreshToken,
@@ -86,8 +83,8 @@ public class RefreshTokenAuth2Provider implements AccessTokenProvider {
     }
   }
 
-  public AccessTokenProvider.AccessToken getAccessToken(
-      String tokenEndpoint,
+  private AccessTokenProvider.AccessToken getAccessToken(
+      String tokenServerUrl,
       String clientId,
       RedactedString clientSecret,
       RedactedString refreshToken,
@@ -96,136 +93,28 @@ public class RefreshTokenAuth2Provider implements AccessTokenProvider {
       RedactedString proxyPassword)
       throws IOException {
 
-    AccessToken token = null;
-    HttpURLConnection conn = null;
     logger.atFine().log("Get a new access token using the refresh token grant flow");
-    try {
-      URL url = new URL(tokenEndpoint);
-      conn = (HttpURLConnection) url.openConnection(getProxy(proxyAddress));
-      conn.setRequestMethod("POST");
-      conn.setReadTimeout(READ_TIMEOUT);
-      conn.setConnectTimeout(CONNECT_TIMEOUT);
-      conn.setRequestProperty("Content-Type", "application/x-www-form-urlencoded");
-      conn.setRequestProperty("Accept", "application/json");
-      if (proxyUsername != null && proxyUsername.value() != null) {
-        logger.atFine().log("Setup the request with the proxy auth");
-        conn.setRequestProperty(
-            HttpHeaders.PROXY_AUTHORIZATION,
-            "Basic "
-                + BaseEncoding.base64()
-                    .encode(
-                        (proxyUsername.value() + ":" + proxyPassword.value())
-                            .getBytes(StandardCharsets.UTF_8)));
-      }
-      conn.setDoOutput(true);
-      String payload =
-          "grant_type=refresh_token&refresh_token="
-              + refreshToken.value()
-              + "&client_id="
-              + clientId
-              + "&client_secret="
-              + clientSecret.value();
-      conn.getOutputStream().write(payload.getBytes("UTF-8"));
+    HttpTransport httpTransport =
+        HttpTransportFactory.createHttpTransport(proxyAddress, proxyUsername, proxyPassword);
+    TokenRequest request =
+        new RefreshTokenRequest(
+                httpTransport, JSON_FACTORY, new GenericUrl(tokenServerUrl), refreshToken.value())
+            .setClientAuthentication(
+                new ClientParametersAuthentication(clientId, clientSecret.value()));
 
-      int httpResponseCode = conn.getResponseCode();
-      logger.atFine().log("Response " + httpResponseCode);
-
-      String responseContentType = conn.getHeaderField("Content-Type");
-
-      if (httpResponseCode == HttpURLConnection.HTTP_OK
-          && responseContentType.startsWith("application/json")) {
-        logger.atFine().log("Received a 200 with presumably a new access token");
-        InputStream httpResponseStream = conn.getInputStream();
-        token = parseTokenFromStream(httpResponseStream);
-      } else {
-        logger.atFine().log("Refresh token grant flow failed");
-        InputStream stream = conn.getErrorStream();
-        if (stream == null) {
-          // no error stream, try the original input stream
-          stream = conn.getInputStream();
-        }
-        String responseBody = consumeInputStream(stream);
-        logger.atSevere().log(
-            "Received following response '"
-                + responseBody
-                + "' with code '"
-                + httpResponseCode
-                + "'");
-      }
-    } finally {
-      if (conn != null) {
-        conn.disconnect();
-      }
-    }
-    return token;
-  }
-
-  private String consumeInputStream(InputStream stream) throws IOException {
-    BufferedReader br = new BufferedReader(new InputStreamReader(stream, "utf-8"));
-    StringBuilder response = new StringBuilder();
-    String responseLine;
-    while ((responseLine = br.readLine()) != null) {
-      response.append(responseLine.trim());
-    }
-    return response.toString();
-  }
-
-  private AccessToken parseTokenFromStream(InputStream httpResponseStream) throws IOException {
-    AccessToken token;
-    try {
-      int expirationPeriodInSecs = 0;
-
-      JsonFactory jf = new JsonFactory();
-      JsonParser jp = jf.createJsonParser(httpResponseStream);
-      String fieldName, fieldValue;
-      jp.nextToken();
-      String accessTokenFromResponse = "";
-      while (jp.hasCurrentToken()) {
-        if (jp.getCurrentToken() == JsonToken.FIELD_NAME) {
-          fieldName = jp.getCurrentName();
-          jp.nextToken(); // field value
-          fieldValue = jp.getText();
-
-          if (fieldName.equals("access_token")) {
-            accessTokenFromResponse = fieldValue;
-          }
-
-          if (fieldName.equals("expires_in")) {
-            expirationPeriodInSecs = Integer.parseInt(fieldValue);
-          }
-        }
-        jp.nextToken();
-      }
-
-      token =
-          new AccessToken(
-              accessTokenFromResponse,
-              System.currentTimeMillis() + (expirationPeriodInSecs * 1000L));
-      jp.close();
-    } finally {
-      httpResponseStream.close();
-    }
-    return token;
-  }
-
-  public Proxy getProxy(String proxyAddress) {
-    if (proxyAddress == null || "".equals(proxyAddress)) {
-      return Proxy.NO_PROXY;
-    }
-    logger.atFine().log("The proxy is defined");
-    String[] proxyAddressParsed = proxyAddress.split(":");
-    return new Proxy(
-        Proxy.Type.HTTP,
-        new InetSocketAddress(proxyAddressParsed[0], Integer.parseInt(proxyAddressParsed[1])));
-  }
-
-  @Override
-  public void setConf(Configuration config) {
-    this.config = config;
+    TokenResponse execute = request.execute();
+    return new AccessToken(
+        execute.getAccessToken(),
+        System.currentTimeMillis() + (execute.getExpiresInSeconds() * 1000L));
   }
 
   @Override
   public Configuration getConf() {
     return this.config;
+  }
+
+  @Override
+  public void setConf(Configuration config) {
+    this.config = config;
   }
 }