+++ util-hadoop/src/main/java/com/google/cloud/hadoop/util/HadoopCredentialConfiguration.java
@@ -132,14 +132,8 @@ public class HadoopCredentialConfiguration {
    * RefreshTokenAuth2Provider
    */
   public static final HadoopConfigurationProperty<String> AUTH_REFRESH_TOKEN_SUFFIX =
-      new HadoopConfigurationProperty<>(".auth.refresh.token");
-
-  /**
-   * Configuration key for defining the OAUth2 token endpoint. Required when the provider is
-   * RefreshTokenAuth2Provider
-   */
-  public static final HadoopConfigurationProperty<String> AUTH_TOKEN_ENDPOINT_SUFFIX =
-      new HadoopConfigurationProperty<>(".auth.token.endpoint");
+      new HadoopConfigurationProperty<>(
+          ".auth.refresh.token", CredentialOptions.TOKEN_SERVER_URL_DEFAULT);
 
   /**
    * Configuration key for defining the OAUth2 client ID. Required when the provider is